feat: Logica del manejo de color, categoria e iconos
