+++ src/teacher/components/benefitsView/benefitCreateView/benefitPreview/BenefitPreview.module.css
@@ -1,0 +1,183 @@
+.previewSection {
+  position: sticky;
+  top: 2rem;
+  height: fit-content;
+}
+
+.previewCard {
+  padding: 1.5rem;
+}
+
+.previewTitle {
+  font-size: 1.125rem;
+  font-weight: 600;
+  color: #2e2e2e;
+  margin-bottom: 1rem;
+  text-align: center;
+}
+
+.benefitPreview {
+  border: 2px solid #e5e7eb;
+  border-radius: 1rem;
+  padding: 1.5rem;
+  background: white;
+}
+
+.previewHeader {
+  display: flex;
+  justify-content: space-between;
+  align-items: flex-start;
+  margin-bottom: 1rem;
+}
+
+.previewBadges {
+  display: flex;
+  gap: 0.5rem;
+  flex-wrap: wrap;
+}
+
+.premiumBadge,
+.limitedBadge {
+  font-size: 0.75rem;
+  padding: 0.25rem 0.5rem;
+  border-radius: 0.375rem;
+  font-weight: 500;
+  display: flex;
+  align-items: center;
+  gap: 0.25rem;
+}
+
+.premiumBadge {
+  background-color: #fef3c7;
+  color: #92400e;
+}
+
+.limitedBadge {
+  background-color: #f3f4f6;
+  color: #374151;
+}
+
+.badgeIcon {
+  font-size: 0.625rem;
+}
+
+.statusBadge {
+  background-color: #d1fae5;
+  color: #065f46;
+  font-size: 0.75rem;
+  padding: 0.25rem 0.75rem;
+  border-radius: 0.375rem;
+  font-weight: 500;
+}
+
+.previewContent {
+  display: flex;
+  align-items: center;
+  gap: 1rem;
+  margin-bottom: 1rem;
+}
+
+.previewIconWrapper {
+  width: 3rem;
+  height: 3rem;
+  border-radius: 0.75rem;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-shrink: 0;
+}
+
+.previewIcon {
+  font-size: 1.25rem;
+  color: white;
+}
+
+.previewInfo {
+  flex: 1;
+}
+
+.previewName {
+  font-size: 1.125rem;
+  font-weight: 700;
+  color: #2e2e2e;
+  margin-bottom: 0.25rem;
+  line-height: 1.2;
+}
+
+.previewCategory {
+  font-size: 0.875rem;
+  color: #007bff;
+  font-weight: 600;
+  text-transform: uppercase;
+  letter-spacing: 0.05em;
+}
+
+.previewDescription {
+  font-size: 0.875rem;
+  color: #6b7280;
+  line-height: 1.5;
+  margin-bottom: 1rem;
+}
+
+.previewStats {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 1rem;
+  padding: 0.75rem;
+  background-color: #f9fafb;
+  border-radius: 0.5rem;
+}
+
+.previewCost {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+}
+
+.previewCostIcon {
+  color: #f59e0b;
+  font-size: 1rem;
+}
+
+.previewDuration {
+  font-size: 0.75rem;
+  color: #6b7280;
+  background-color: #f3f4f6;
+  padding: 0.25rem 0.5rem;
+  border-radius: 0.25rem;
+  font-weight: 500;
+}
+
+.previewRestrictions {
+  margin-top: 1rem;
+}
+
+.previewRestrictionsTitle {
+  font-size: 0.75rem;
+  font-weight: 600;
+  color: #374151;
+  display: block;
+  margin-bottom: 0.5rem;
+}
+
+.previewRestrictionsList {
+  list-style: none;
+  padding: 0;
+  margin: 0;
+}
+
+.previewRestrictionsList li {
+  font-size: 0.75rem;
+  color: #6b7280;
+  padding: 0.125rem 0;
+  position: relative;
+  padding-left: 1rem;
+}
+
+.previewRestrictionsList li::before {
+  content: "•";
+  position: absolute;
+  left: 0;
+  color: #d1d5db;
+}
