+++ src/teacher/components/benefitsView/benefitCreateView/benefitRestrictions/BenefitRestrictions.module.css
@@ -1,0 +1,46 @@
+.section {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+}
+
+.sectionTitle {
+  font-size: 1.125rem;
+  font-weight: 600;
+  color: #2e2e2e;
+  margin-bottom: 0.5rem;
+  padding-bottom: 0.5rem;
+  border-bottom: 2px solid #f3f4f6;
+}
+.restrictionsSection {
+  display: flex;
+  flex-direction: column;
+  gap: 0.75rem;
+}
+
+.restrictionItem {
+  display: flex;
+  gap: 0.5rem;
+  align-items: center;
+}
+
+.restrictionItem input {
+  flex: 1;
+}
+
+.removeButton {
+  color: #ef4444;
+  padding: 0.5rem;
+  min-width: auto;
+}
+
+.removeButton:hover {
+  background-color: #fef2f2;
+  color: #dc2626;
+}
+
+.addButton {
+  align-self: flex-start;
+  font-size: 0.875rem;
+  padding: 0.5rem 1rem;
+}
