+++ src/teacher/components/benefitsView/benefitCreateView/benefitRestrictions/BenefitRestrictions.tsx
@@ -1,0 +1,48 @@
+// import styles from "./BenefitRestrictions.module.css";
+
+// IMPLEMENTAR EN PROXIMO SPRINT
+const BenefitRestrictions = () => {
+  return {
+    /* <div className={styles.section}>
+                <h3 className={styles.sectionTitle}>
+                  Restricciones y Condiciones
+                </h3>
+                <div className={styles.restrictionsSection}>
+                  {formData.restrictions.map((restriction, index) => (
+                    <div key={index} className={styles.restrictionItem}>
+                      <Input
+                        placeholder="Ej: Solo aplicable a parciales"
+                        value={restriction}
+                        onChange={(e) =>
+                          updateRestriction(index, e.target.value)
+                        }
+                      />
+                      {formData.restrictions.length > 1 && (
+                        <Button
+                          type="button"
+                          variant="ghost"
+                          size="sm"
+                          onClick={() => removeRestriction(index)}
+                          className={styles.removeButton}
+                        >
+                          <FaTrash />
+                        </Button>
+                      )}
+                    </div>
+                  ))}
+                  <Button
+                    type="button"
+                    variant="outline"
+                    size="sm"
+                    onClick={addRestriction}
+                    className={styles.addButton}
+                  >
+                    <FaPlus className={styles.buttonIcon} />
+                    Agregar Restricción
+                  </Button>
+                </div>
+              </div> */
+  };
+};
+
+export default BenefitRestrictions;
