+++ src/teacher/contexts/benefitsAPIContext/BenefitAPIContext.ts
@@ -1,0 +1,6 @@
+import { createContext } from "react";
+import type { BenefitAPIContextType } from "./BenefitAPIContext.type";
+
+export const BenefitAPIContext = createContext<
+  BenefitAPIContextType | undefined
+>(undefined);
