--- src/teacher/contexts/benefitsContext/BenefitContext.ts
@@ -1,6 +1,0 @@
-import { createContext } from "react";
-import type { BenefitContextType } from "./BenefitContext.type";
-
-export const BenefitContext = createContext<BenefitContextType | undefined>(
-  undefined
-);
