+++ src/teacher/contexts/benefitsUIContext/BenefitUIContext.ts
@@ -1,0 +1,6 @@
+import { createContext } from "react";
+import type { BenefitUIContextType } from "./BenefitUIContext.type";
+
+export const BenefitUIContext = createContext<BenefitUIContextType | undefined>(
+  undefined
+);
