+++ src/teacher/views/benefitsView/benefitsViewList/BenefitsListView.module.css
@@ -51,6 +51,7 @@
   grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
   gap: 1.5rem;
   margin-bottom: 2rem;
+  max-width: 300px;
 }
 
 .statCard {
